requirements.txt
pillow
opencv-python
pyserial
numpy
piexif
matplotlib
pandas
exifread
pyexiv2
